+++ CHANGELOG.md
@@ -61,6 +61,8 @@
   (and the grid reflowed) or not when e.g. zooming in/out
   ([#1807][1807]).
 * `strikeout-thickness` option.
+* Implemented the new `xdg-toplevel-icon-v1` protocol.
+* Implemented `CSI 21 t`: report window title.
 
 [1807]: https://codeberg.org/dnkl/foot/issues/1807